Fix `yulcode` Task (#49)
This PR fixes the existing `yulcode` task. The test it was using was not
strict enough, in particular `npx hardhat yulcode --contractname
SafeTokenLock` would produce output for both `ISafeTokenLock` and
`SafeTokenLock`, which was incorrect. Additionally, it would dump a JSON
which is formatted by NodeJS and doesn't print things nicely (in
particular, the Yul code is escaped and hard to read, and many of the
nested properties are output as `[Object]` or `[Array]` instead of their
values).

In addition, errors are propagated correctly by setting the
`process.errorCode`.

I noticed this when trying to understand why `unlock` and `withdraw` was
so significantly more expensive in #48.