Add view methods (#53)
Fixes #16 
Changes in PR:

- mark storage variables `users` and `unlocks` as `internal` so that
compiler does not generate getter functions
- Add two view functions:
    - Get `User` struct value for given address
    - Get `unlock` information for given address and index
- Fix breaking tests

Why this approach?

a. In the current contract implementation, the state variables are
already public meaning, compiler would generate getter functions for
them and contract size is:
`SafeTokenLock 4920 bytes (limit is 24576)`
b. whereas, with the changes in [this
commit](https://github.com/safe-global/safe-locking/commit/8dbdb585e28e7048e6ea94253daa613f739b6258)
codesize increases as follows which is not a problem for Locking
contract:
`SafeTokenLock 5093 bytes (limit is 24576)`
c. and with a, b combined i.e. public variable with additional getter
functions
`SafeTokenLock 5498 bytes (limit is 24576)`
I do not prefer this approach because we have additional unnecessary
getter functions.


I would go for `b` as it would help encapsulate details of how we store
the user related information in the storage variables as already stated
in the issue. Let me know if you have contradictory opinion.


The code coverage is already 100% and already covers expected output
assertions in the tests so no new tests added.

---------

Co-authored-by: Shebin John <admin@remedcu.com>
Co-authored-by: Mikhail <16622558+mmv08@users.noreply.github.com>