Add Immutables To Interface (#62)
Fixes #61, #57 

This PR adds the Safe token address and cooldown period to the
`ISafeTokenLock` interface.

I noticed that the tests for unlocking were taking the current timestamp
instead of the block timestamp from the `unlock` transaction, so I
changed those tests to be a bit more explicit about which timestamp is
important (although, in practice there is no difference). This allows us
to close #57 as invalid, as the test shows that the bug is actually not
present :).