Document Max Unlocks Limitation (#76)
Resolves #59 

This PR documents the limitation that a maximum of unlocks that a single
user can perform. In particular, after 4294967296 unlocks, the
`unlockEnd` user field will begin to overflow and cause `unlock`
transactions to revert.

However, this is not really an issue for users, seeing as performing
that many unlocks is not really feasible. In particular, we can perform
a maximum of ~1167 unlocks in a single block (see the newly added test;
give or take a few calls depending on what optimizations are possible to
make). This means that you can, at most, unlock with an average of `30e6
/ 1167 ~= 25700` gas per unlock. This would require a total of
`110380659507200` total gas (110 trillion) to reach. Assuming current
gas prices of around 30 Gwei and without considering the fact that the
EIP-1559 block base gas fee would increase as a result of full blocks,
this would result in around 3.3 million Ether required to execute these
transactions.