Add Safe Token Invariants (#77)
This PR adds a new Safe token spec file to prove important invariants
required for the locking contract formal verification.

Note that, because of limitations of the Certora tool, we cannot include
these files directly in the existing `SafeTokenLock.spec` file (as
invariants can only be defined for the current contract). Instead, in
the Safe token lock spec, we add a `setup-` function that `require`s the
equivalent of the invariants that are proven in the Safe token spec.