Implement Unlock Correctness Rules (#78)
This PR implements correctness rules for the `unlock` function. Namely,
it proves that it is always possible to unlock your locked tokens (a
nice property to have as a user!).

Note that this requires proving some additional invariants to make this
possible. Notably that:
- The sum of all locked and unlocked tokens is less than the total Safe
token balance in the locking contract (important to prevent overflowing
the `unlock` amounts when proving the above rule)
- Unlocks start index is always before the unlock end index (so adding
new unlocks works as expected - since it is added to the back of the
list instead of somewhere that is not "seen")
- Unlock amounts are non-zero
- The user unlocked amount is equal to the individual unlock amounts
(also important for overflow reasons).

All in all, a nice rabbit hole, but finally found an end to it :smile:.