Formally Verify `withdraw` Commutativity (#79)
This PR adds a rule to verify withdrawal commutativity. This means that
regardless of the order you withdraw, as long as you use the same
`maxUnlocks` values you will end in the same result.

Note that there was previously an exception to commutativity when
`maxUnlocks2 == 0`, as the second withdraw could overflow
`user.unlockEnd + maxUnlocks1` and cause it to revert. I slightly
reworked the `withdraw` function so this is no longer possible, and
saved around 60 gas to boot :tada:. With this change, `withdraw` is
fully commutative as well, with no exceptions.

I'm very happy overall, FV lead to code improvement :smile:.