Timestamp increase verification for Unlock operation (#81)
This PR contains the timestamp-based verification to see if the
timestamp is increasing with each unlock operations. Also, checks if
there are two or more unlock operations for a particular user, then the
newer unlock operation maturity time is always higher than the older
unlock operations.

Also added a rule which verifies that the `SAFE_TOKEN` and
`COOLDOWN_PERIOD` is never changed.