Use Unchecked Math for Unlock Maturity Timestamp Computation (#83)
This PR changes the computation of the `maturesAt` timestamp to be
unchecked. The exact reason is documented in the contract, but it boils
down to:
- We will not reach this condition in our lifetime
- Has the nice property that, if someone has tokens in the Safe token
lock contract hundreds of billions of years after the Sun went
supernova, they will be able to withdraw them immediately. I mean, they
earned it locking Safe for so long!