Explicitely Revert On Invalid Total Supply (#87)
This PR slightly changes the constructor logic to require that the token
that is passed in has a total supply that fits in a `uint96`. This
change was done since:
1. The contract assumes this to be true for it to function correctly. I
thought it would be good to document it explicitly as a condition that
is verified on-chain
2. The gas cost overhead of the check is only payed once on deployment
3. We still revert when constructing the a locking contract instance
with `address(0)`.

Added a new unit test to verify new logic.